объединены модули по работе с модальными окнами
